+++ internal/model/message/deleted.go
@@ -23,7 +23,14 @@ func CreateDeleted(db *gorm.DB, d *Deleted) error {
 	return db.Create(d).Error
 }
 
-func UpdateDeleted(db *gorm.DB, d *Deleted) error {
+func UpdateDeletedById(db *gorm.DB, d *Deleted, id int) error {
+	var deleted Deleted
+	result := db.First(&deleted, id)
+
+	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
+		return DeletedNotFoundErr
+	}
+
 	return db.Save(d).Error
 }
 
@@ -42,5 +49,12 @@ func GetDeletedById(db *gorm.DB, id int) (Deleted, error) {
 }
 
 func DeleteDeletedById(db *gorm.DB, id int) error {
+	var deleted Deleted
+	result := db.First(&deleted, id)
+
+	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
+		return DeletedNotFoundErr
+	}
+
 	return db.Delete(&Deleted{}, id).Error
 }
